requirements.txt
asgiref==3.10.0
Django==5.2.8
psycopg==3.2.12
psycopg-binary==3.2.12
sqlparse==0.5.3
django-smart-selects==1.7.2